% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{extractGenericJuice}
\alias{extractGenericJuice}
\title{extractGenericJuice}
\usage{
extractGenericJuice(X_train, Y_train, numFolds, parCV, numCores, seedNum,
  verbose_p, fn, fn_params)
}
\arguments{
\item{X_train}{Training Data (excludes the response/target we wish to predict
) that will be fed into the pipeline function.}

\item{Y_train}{Training Response/Target - The response/target that will be fed into the pipeline function.}

\item{numFolds}{Integer indicating the number of folds to use to extract data}

\item{parCV}{Boolean indicating whether to parallelize the extraction prodcedure.}

\item{numCores}{Integer indicating the number of cores to use when generating predictions.}

\item{seedNum}{Integer indicating the seed number. Using the same seed will generate the same folds.}

\item{verbose_p}{Boolean indicating if prediction extraction details should be printed out the screen.}

\item{fn}{The pipeline function. The pipeline function must take parameters training data, training response,
validation data, validation response. See examples for details.}

\item{fn_params}{Additional parameters to supply to the pipeline function. See examples for details.}
}
\value{
CV Fold Predictions
}
\description{
Returns cross validation data from folds (i.e. if 2 folds, then aggregate the data on
fold 2 after training only on fold 1 & the data on fold 1 after training only on fold 2). This
is helps in feature engineering.
}

